add user-profile-form components
